requirements.txt
streamlit>=1.32.0
openai>=1.14.0
python-docx>=1.1.0
pypdf>=4.1.0
requests>=2.31.0
python-dotenv>=1.0.0
